Add functional tests for admin adding tag categories and journalists using tags
Tests verify that an admin is able to add and remove tag categories from the
SecureDrop server, that a journalist is able to add and remove labels from a particular
source or submission, and that they are able to filter sources by particular labels.

This also modifies the templates to add ids to some elements for ease of testing.